requirements.txt
flask
google-genai
pip-system-certs